# Implement From float specialization for PluralOperands