Based on the suggestion from https://users.rust-lang.org/u/alice in https://users.rust-lang.org/t/get-fractionals-out-of-f64-as-integer/37559/8