@zbraniecki 2020-01-29
Perf:
without: parse_float_operands time: [1.1371 us 1.1428 us 1.1501 us]
with: parse_float_operands time: [101.18 ns 101.38 ns 101.62 ns]