@zbraniecki 2020-01-29
@Manishearth - does it look sane to you?
Do you think we should keep signed integers in TryFrom, and unsigned/float in From, or would it be ok to move signed integers to From as well (and switch from `checked_abs`)?