@Manishearth 2020-01-29
Yeah I think it would be fine.