@zbraniecki 2020-05-27
Possible crate to use - https://github.com/Alexhuszagh/rust-lexical/blob/master/README.md#benchmarks